feat: add `parser` module, a bit more logic [ci skip]
